tmp/chatgpt/requirements.txt
pytube==12.1.2
python-dotenv==0.21.1
openai==0.26.5